requirements.txt
Flask==2.3.3
PyJWT==2.8.0
Werkzeug==2.3.7
psycopg2-binary>=2.9.0  # Cambiar a versión flexible
bcrypt==4.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23      # Alternativa opcional